Refactor: Split main.go into multiple files
This commit refactors the monolithic `main.go` file into a more organized and modular structure. The original file has been split into multiple smaller files, which are grouped by their logical function into subdirectories under the `src/` directory.

The new structure is as follows:
- `src/main.go`: Contains the `main` function, the entry point of the application.
- `src/api/`: Contains files related to API interaction.
- `src/config/`: Contains files for configuration and settings.
- `src/conversation/`: Contains files for managing conversation history.
- `src/ui/`: Contains files for user interface elements, such as help text and terminal colors.
- `src/utils/`: Contains utility functions.

A `build.sh` script has been added to simplify the build process for the new file structure. This script compiles all the `.go` files into a single executable named `nvidia-chat-refactored`.

This refactoring improves the maintainability and readability of the code without changing its functionality.
