+++ src/utils/files.go
@@ -1,0 +1,32 @@
+// Path: src/utils/files.go
+// Rationale: This file contains utility functions for file operations.
+
+package main
+
+import (
+	"io"
+	"os"
+)
+
+func copyFile(src, dst string) error {
+	in, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer in.Close()
+	out, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+	_, err = io.Copy(out, in)
+	if err != nil {
+		return err
+	}
+	return out.Sync()
+}
+
+func fileExists(path string) bool {
+	_, err := os.Stat(path)
+	return err == nil
+}
